Fully functional pipeline ready

+++ src/Shingler.cpp
@@ -3,8 +3,9 @@
 std::set<std::string> Shingler::get_shingles(const std::string& text, int k) 
 {
     std::set<std::string> shingles;
-    if (text.length() < k) return shingles;
-    for (size_t i = 0; i <= text.length() - k; ++i) {
+    if (text.length() < static_cast<size_t>(k)) return shingles;
+    for (size_t i = 0; i <= text.length() - k; ++i) 
+    {
         shingles.insert(text.substr(i, k));
     }
     return shingles;
